binding-core.c: add cast to remove warnings; ref #5362

+++ mrbgems/mruby-binding-core/src/binding-core.c
@@ -198,8 +198,8 @@ binding_source_location(mrb_state *mrb, mrb_value self)
       srcloc = mrb_nil_value();
     }
     else {
-      const char *fname = mrb_debug_get_filename(mrb, irep, pc);
-      mrb_int fline = mrb_debug_get_line(mrb, irep, pc);
+      const char *fname = mrb_debug_get_filename(mrb, irep, (uint32_t)pc);
+      mrb_int fline = mrb_debug_get_line(mrb, irep, (uint32_t)pc);
 
       if (fname && fline >= 0) {
         srcloc = mrb_assoc_new(mrb, mrb_str_new_cstr(mrb, fname), mrb_fixnum_value(fline));
